refactor: 정보탭 관련 컴포넌트 파일 분리
