+++ src/components/Info/CategoryToggle.tsx
